+++ shenyu-examples/shenyu-examples-sdk/shenyu-examples-sdk-dubbo/shenyu-examples-sdk-apache-dubbo-provider/pom.xml
@@ -81,7 +81,7 @@
         <dependency>
             <groupId>org.apache.shenyu</groupId>
             <artifactId>shenyu-examples-apache-dubbo-service-xml</artifactId>
-            <version>2.5.1-SNAPSHOT</version>
+            <version>${project.version}</version>
         </dependency>
 
         <!-- Keep latest Nacos client version -->
