+++ shenyu-integrated-test/pom.xml
@@ -41,6 +41,7 @@
         <module>shenyu-integrated-test-websocket</module>
         <module>shenyu-integrated-test-combination</module>
         <module>shenyu-integrated-test-sdk-apache-dubbo</module>
+        <module>shenyu-integrated-test-sdk-alibaba-dubbo</module>
     </modules>
 
     <properties>
